# backend/docs/specs/Endpoints_OpenAPI_Spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task Management API",
    "version": "1.0.0",
    "description": "OpenAPI 3.0 specification for a task management system with authentication, task CRUD, user management, and error handling."
  },
  "servers": [
    { "url": "/api/v1" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "due_date": { "type": "string", "format": "date" },
          "completed": { "type": "boolean" },
          "priority": { "type": "string" },
          "category": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "code": { "type": "integer" }
        }
      }
    }
  },
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/tasks": {
      "get": {
        "summary": "Get all tasks for the authenticated user",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["all", "completed", "incomplete"] } },
          { "name": "priority", "in": "query", "schema": { "type": "string" } },
          { "name": "due", "in": "query", "schema": { "type": "string", "enum": ["today", "this_week", "overdue"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of tasks", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Task" } } } } }
        }
      },
      "post": {
        "summary": "Create a new task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "priority": { "type": "string" },
                  "category": { "type": "string" },
                  "due_date": { "type": "string", "format": "date" }
                },
                "required": ["title"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Task created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Task" } } } },
          "400": { "description": "Invalid input", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}